+++ docs/CHANGELOG.md
@@ -23,6 +23,7 @@ General:
    - Fixed volume merging issues
    - Added Quake2 Model (`md2`) support
    - Removed cvar `voxformat_marchingcubes` and replaced with `voxel_meshmode` (set to `1` to use marching cubes)
+   - Added new lua script `gradient.lua`
 
 VoxConvert: